Move files into appropriate folders
